fix: move verification service to role module
