+++ docs/api/modules.md
@@ -69,7 +69,6 @@
 - [modules/claim/services](modules/modules_claim_services.md)
 - [modules/claim/services/claim-issuance.service](modules/modules_claim_services_claim_issuance_service.md)
 - [modules/claim/services/claim.service](modules/modules_claim_services_claim_service.md)
-- [modules/claim/services/issuer-verification.service](modules/modules_claim_services_issuer_verification_service.md)
 - [modules/claim/services/revocation-verification.service](modules/modules_claim_services_revocation_verification_service.md)
 - [modules/did-contact/did.contact.controller](modules/modules_did_contact_did_contact_controller.md)
 - [modules/did-contact/did.contact.dto](modules/modules_did_contact_did_contact_dto.md)
@@ -104,6 +103,7 @@
 - [modules/organization/organization.module](modules/modules_organization_organization_module.md)
 - [modules/organization/organization.resolver](modules/modules_organization_organization_resolver.md)
 - [modules/organization/organization.service](modules/modules_organization_organization_service.md)
+- [modules/role/issuer-verification.service](modules/modules_role_issuer_verification_service.md)
 - [modules/role/role.controller](modules/modules_role_role_controller.md)
 - [modules/role/role.dto](modules/modules_role_role_dto.md)
 - [modules/role/role.entity](modules/modules_role_role_entity.md)
