+++ docs/api/modules/modules_claim_services.md
@@ -6,7 +6,6 @@
 
 - [ClaimIssuanceService](modules_claim_services.md#claimissuanceservice)
 - [ClaimService](modules_claim_services.md#claimservice)
-- [IssuerVerificationService](modules_claim_services.md#issuerverificationservice)
 - [RevocationVerificationService](modules_claim_services.md#revocationverificationservice)
 
 ## References
@@ -20,12 +19,6 @@ ___
 ### ClaimService
 
 Re-exports [ClaimService](../classes/modules_claim_services_claim_service.ClaimService.md)
-
-___
-
-### IssuerVerificationService
-
-Re-exports [IssuerVerificationService](../classes/modules_claim_services_issuer_verification_service.IssuerVerificationService.md)
 
 ___
