--- docs/api/modules/modules_claim_services_issuer_verification_service.md
@@ -1,7 +1,0 @@
-# Module: modules/claim/services/issuer-verification.service
-
-## Table of contents
-
-### Classes
-
-- [IssuerVerificationService](../classes/modules_claim_services_issuer_verification_service.IssuerVerificationService.md)
