+++ docs/api/modules/modules_role_issuer_verification_service.md
@@ -1,0 +1,7 @@
+# Module: modules/role/issuer-verification.service
+
+## Table of contents
+
+### Classes
+
+- [IssuerVerificationService](../classes/modules_role_issuer_verification_service.IssuerVerificationService.md)
