+++ src/modules/claim/claim.module.ts
@@ -10,7 +10,6 @@ import {
   ClaimService,
   ClaimIssuanceService,
   RevocationVerificationService,
-  IssuerVerificationService,
 } from './services';
 import { AssetsModule } from '../assets/assets.module';
 import { DIDModule } from '../did/did.module';
@@ -33,12 +32,7 @@ import { RoleRevokerResolver } from './resolvers/revoker.resolver';
     RevocationVerificationService,
     Provider,
     RoleRevokerResolver,
-    IssuerVerificationService,
   ],
-  exports: [
-    RevocationVerificationService,
-    RoleRevokerResolver,
-    IssuerVerificationService,
-  ],
+  exports: [RevocationVerificationService, RoleRevokerResolver],
 })
 export class ClaimModule {}
