+++ src/modules/claim/services/claim.service.ts
@@ -80,7 +80,7 @@ export class ClaimService {
       claimType,
       claimTypeVersion,
     });
-
+    //here is verify
     await this.roleService.verifyEnrolmentPrecondition({
       claimType,
       userDID: dto.requester,
