+++ src/modules/claim/services/index.ts
@@ -1,4 +1,3 @@
 export * from './claim-issuance.service';
 export * from './claim.service';
 export * from './revocation-verification.service';
-export * from './issuer-verification.service';
