+++ src/modules/role/issuer-verification.service.ts
@@ -1,13 +1,13 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { verifyCredential } from 'didkit-wasm-node';
 import { RegistrySettings } from '@ew-did-registry/did-resolver-interface';
-import { Provider } from '../../../common/provider';
-import { RoleService } from '../../role/role.service';
-import { RoleCredentialResolver } from '../resolvers/credential.resolver';
-import { RoleIssuerResolver } from '../resolvers/issuer.resolver';
-import { DIDService } from '../../did/did.service';
+import { Provider } from '../../common/provider';
+import { RoleService } from './role.service';
+import { RoleCredentialResolver } from '../claim/resolvers/credential.resolver';
+import { RoleIssuerResolver } from '../claim/resolvers/issuer.resolver';
+import { DIDService } from '../did/did.service';
 import { IssuerVerification } from '@energyweb/vc-verification';
-import { RevocationVerificationService } from './revocation-verification.service';
+import { RevocationVerificationService } from '../claim/services/revocation-verification.service';
 
 @Injectable()
 export class IssuerVerificationService extends IssuerVerification {
